+++ include/libpbl/pbl.h
@@ -9,9 +9,15 @@
 #pragma once
 
 // Including Headers that are required throughout the entire PBL Library
-#include <stdarg.h>
-#include <stdio.h>
-#include <stdlib.h>
+#ifdef __cplusplus
+  #include <cstdarg>
+  #include <cstdio>
+  #include <cstdlib>
+#else
+  #include <stdarg.h>
+  #include <stdio.h>
+  #include <stdlib.h>
+#endif
 
 // Main Header Inclusions
 #include "./pbl-main.h"